pyproject.toml
[project]
name = "hackspark"
version = "0.1.0"
description = "Hackathon-ready FastAPI microservices boilerplate."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite==0.21.0",
    "alembic==1.16.4",
    "argon2-cffi==25.1.0",
    "asyncpg==0.30.0",
    "email-validator==2.2.0",
    "fastapi==0.116.1",
    "httpx==0.28.1",
    "prometheus-client==0.23.1",
    "pydantic-settings==2.10.1",
    "python-jose[cryptography]==3.5.0",
    "sqlalchemy[asyncio]==2.0.43",
    "structlog==25.4.0",
    "uvicorn[standard]==0.35.0",
]

[dependency-groups]
dev = [
    "pytest==8.4.2",
    "ruff==0.13.3",
    "ty==0.0.1a19",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["tests"]
markers = [
    "integration: end-to-end tests that span multiple services (slow; skip with -m 'not integration')",
    "unit: fast isolated unit tests",
]

[tool.ty.environment]
python-version = "3.12"
extra-paths = [
    ".",
    "gateway",
    "services/auth_service",
    "services/item-service",
    "services/ai_agent_service",
]

[tool.ty.terminal]
error-on-warning = true
output-format = "concise"
